serviços e controllers basicos

+++ src/main/java/br/ufrn/imd/gourmetize_backend/repository/ReceitaRepository.java
@@ -3,5 +3,5 @@ package br.ufrn.imd.gourmetize_backend.repository;
 import br.ufrn.imd.gourmetize_backend.model.Receita;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReceitaRepository extends JpaRepository<Receita, Integer> {
+public interface ReceitaRepository extends JpaRepository<Receita, Long> {
 }
